refactor: Restructure folder structure for RepoList
