+++ src/pages/RepoList/index.tsx
@@ -2,11 +2,11 @@ import { useEffect, useState } from "react";
 import { Helmet } from "react-helmet";
 import { useNavigate } from "react-router-dom";
 import { useGetRepoQuery } from "../../state/issueRTK";
-import Footer from "../Footer";
-import Header from "../Header";
+import Footer from "../../component/Footer";
+import Header from "../../component/Header";
 import RepoLayout from "./RepoLayout";
 
-function RepoWrapper() {
+const RepoList = () => {
   const [userInfo, setUserInfo] = useState<any>();
   const [skip, setSkip] = useState(true);
   const navigate = useNavigate();
@@ -66,4 +66,4 @@ function RepoWrapper() {
   );
 }
 
-export default RepoWrapper;
+export default RepoList;
